fix: Double->long cast

+++ src/main/java/com/azeluxclient/module/render/TimeChanger.java
@@ -5,7 +5,6 @@ import com.azeluxclient.setting.SliderSetting;
 import net.minecraft.client.MinecraftClient;
 
 public class TimeChanger extends Module {
-    // 0 = dawn, 6000 = noon, 12000 = dusk, 18000 = midnight
     private final SliderSetting time = register(new SliderSetting("Time", 6000.0, 0.0, 24000.0));
     private long savedTime = -1;
 
@@ -29,6 +28,7 @@ public class TimeChanger extends Module {
     @Override
     public void onTick(MinecraftClient client) {
         if (client.world == null) return;
-        client.world.getLevelProperties().setTimeOfDay((long) time.getValue());
+        // Fix: explicit double unbox before long cast
+        client.world.getLevelProperties().setTimeOfDay((long)(double) time.getValue());
     }
 }
